move around tokenizer things
